create unit tests in the users module and remove the route configuration in tsconfig.json

+++ src/modules/users/models/tokenJwt.model.ts
@@ -11,7 +11,7 @@ export class RefreshToken extends Model {
     @ForeignKey(() => User)
     @Column({
         primaryKey: true,
-        type: DataType.INTEGER,
+        type: DataType.BIGINT,
         autoIncrement: true,
         field: "id_refresh_token"
     })
